Finished the Auth scheme. Users can now be created, login, create a topic, post a new comment and delete their post/topic. All with 2 level Mandatory Access Control.

+++ src/backend/routes/userRoutes.js
@@ -46,12 +46,12 @@ userRoutes.get("/public", function(req,res){
 })
 
 //APP_USE only: requires that the user is Authenticated
-userRoutes.get("/authenticated", function(req,res){
+userRoutes.get("/authenticated", requireAuth, function(req,res){
   findDocs({}, "-__v -password", userModel, userFindCallback(req, res, "Returning all users"))
 })
 
 //Get a user by userId
-.get("/:userId", function(req, res){
+.get("/:userId", requireAuth, function(req, res){
   findOneDoc({_id:req.params.userId}, "-__v -_id -password", userModel, userFindCallback(req, res, "Found user"))
 })
 
@@ -77,25 +77,27 @@ userRoutes.get("/authenticated", function(req,res){
     })
 })*/
 
-.delete("/:userId", function(req, res){
-  deleteDoc(req.params.userId, userModel, userDeleteCallback(req, res, null))
+.delete("/:userId", requireAuth, function(req, res){
+  if(req.user._id == req.params.userId || req.user.role == ROLE_ADMIN)
+    deleteDoc(req.params.userId, userModel, userDeleteCallback(req, res, null))
 })
 
 //Updates a certain user properties given a certain user Id
 //Example usage: Editing a user profile given that the user has been authenticated
-.put("/:userId",  function(req, res){
+.put("/:userId",  requireAuth, function(req, res, next){
+  if(req.user._id == req.params.userId)
+    updateDoc(req.params.userId, { role : ROLE_ADMIN }, {  _id : 0, password : 0, __v : 0 }, userModel, userUpdateCallback(req, res, null))
+  else
+    next()
+}, roleAuth(ROLE_ADMIN), function(req,res){
   updateDoc(req.params.userId, { role : ROLE_ADMIN }, {  _id : 0, password : 0, __v : 0 }, userModel, userUpdateCallback(req, res, null))
 })
 
-.put("/:userId/setToAdmin", function(req, res){
-  updateDoc(req.params.userId, { role : ROLE_ADMIN }, { role : 1, _id : 0 }, userModel, userUpdateCallback(req, res, "User successfully promoted to Admin"))
+.put("/:userId/setToAdmin", requireAuth, roleAuth(ROLE_ADMIN), function(req, res){
+    updateDoc(req.params.userId, { role : ROLE_ADMIN }, { username : 1, _id : 0 }, userModel, userUpdateCallback(req, res, "User successfully promoted to Admin"))
 })
 
-.put("/:userId/setToMod", function(req, res){
-  updateDoc(req.params.userId, { role : ROLE_MODERATOR }, { role : 1, _id : 0 }, userModel, userUpdateCallback(req, res, "User successfully promoted to Moderator"))
-})
-
-.put("/:userId/setToUser", function(req, res){
+.put("/:userId/setToUser", requireAuth, roleAuth(ROLE_ADMIN), function(req, res){
   updateDoc(req.params.userId, { role : ROLE_USER }, { role : 1, _id : 0 }, userModel, userUpdateCallback(req, res, "User successfully promoted to Moderator"))
 })
